requirements.txt
alembic==1.13.1
aniso8601==9.0.1
async-timeout==4.0.3
attrs==23.2.0
awscli==1.32.113
bcrypt==4.1.2
blinker==1.7.0
boto3==1.34.113
botocore==1.34.113
cachelib==0.9.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
docutils==0.16
et-xmlfile==1.1.0
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Caching==2.2.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Mail==0.9.1
Flask-Migrate==4.0.7
Flask-Mpesa==1.2.8
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.7
importlib_resources==6.4.0
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Mako==1.3.2
MarkupSafe==2.1.5
mysql-connector-python==8.4.0
mysqlclient==2.2.4
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.2
pyasn1==0.6.0
PyJWT==2.8.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
referencing==0.34.0
requests==2.32.0
rpds-py==0.18.0
rsa==4.7.2
s3transfer==0.10.1
six==1.16.0
SQLAlchemy==2.0.29
SQLAlchemy-serializer==1.4.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.1